refactor: BlackJackGame -> game 변경
