+++ src/main/java/domain/game/BlackJackAction.java
@@ -1,4 +1,4 @@
-package domain.BlackJackGame;
+package domain.game;
 
 public enum BlackJackAction {
     HIT, STAND;
